libgo: fix build on AIX
    Since aix/ppc64 has been added to GC toolchain, a mix between new and
    old files were created in gcc toolchain.
    This commit corrects this merge for aix/ppc64 and aix/ppc.
    
    Reviewed-on: https://go-review.googlesource.com/c/gofrontend/+/167658

From-SVN: r269797

+++ libgo/go/cmd/go/internal/lockedfile/internal/filelock/filelock_fcntl.go
@@ -35,7 +35,9 @@ const (
 	writeLock lockType = syscall.F_WRLCK
 )
 
-type inode = uint64 // type of syscall.Stat_t.Ino
+// type of syscall.Stat_t.Ino for 64 bits architectures.
+// For 32 bits architecture, it's easier to cast it instead.
+type inode = uint64
 
 type inodeLock struct {
 	owner File
@@ -59,7 +61,7 @@ func lock(f File, lt lockType) (err error) {
 	if err != nil {
 		return err
 	}
-	ino := fi.Sys().(*syscall.Stat_t).Ino
+	ino := uint64(fi.Sys().(*syscall.Stat_t).Ino)
 
 	mu.Lock()
 	if i, dup := inodes[f]; dup && i != ino {
